renegade-metrics: remove dependency on state module
