+++ renegade-metrics/Cargo.toml
@@ -19,4 +19,3 @@ num-bigint = { workspace = true }
 circuit-types = { path = "../circuit-types" }
 util = { path = "../util" }
 common = { path = "../common" }
-state = { path = "../state" }
