+++ workers/gossip-server/src/peer_discovery/heartbeat.rs
@@ -13,11 +13,10 @@ use gossip_api::{
     },
 };
 use job_types::network_manager::{NetworkManagerControlSignal, NetworkManagerJob};
-use renegade_metrics::helpers::record_num_peers_metrics;
 use tracing::info;
 use util::{err_str, get_current_time_millis};
 
-use crate::{errors::GossipError, server::GossipProtocolExecutor};
+use crate::{errors::GossipError, peer_discovery::peer_metrics::record_num_peers_metrics, server::GossipProtocolExecutor};
 
 // -------------
 // | Constants |
