+++ workers/gossip-server/src/peer_discovery/mod.rs
@@ -1,6 +1,7 @@
 //! Groups handlers for peer discovery and indexing
 
 pub(crate) mod expiry_window;
+pub(crate) mod peer_metrics;
 pub mod heartbeat;
 pub mod heartbeat_timer;
 pub mod peers;
